PV-654 Rework auth process to fix several issues
If getUser() returns a falsy value it is assumed they are not
authenticated by the parent application and they are redirected to /

The parent application should also specify a isEnabledForUser() closure
in their config. This is used to determine whether to apply 2FA or not.
previously this required getUser() to return false, but this is not a
reliable way to check (the absence of a username is also false).

Aside from showing the register and authentication pages, all responses
return json data rather than a HTML page if it was a xhr request.

If the initial request is a xhr request, a 401 response is returned
rather than a redirect.

I did away with storing the authenticator ID in the user's session. It
wasn't used anywhere so finding the authenticator by username is
required anyway. The useSession property is no longer required so has
been removed.

It is now possible to specify the expiry time in minutes. If set it will
be used, otherwise the expireDays property will be used instead.